requirements.txt
flet
pytube
openai-whisper
